Django Api/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2